Start a new file for the query subsystem
