+++ opencog/persist/api/CMakeLists.txt
@@ -1,5 +1,6 @@
 
 ADD_LIBRARY (persist
+	BackingQuery.cc
 	BackingStore.cc
 	PersistSCM.cc
 )
